@@ -56,20 +56,21 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.1.3",
-    "@angular-eslint/builder": "14.0.0-alpha.3",
-    "@angular-eslint/eslint-plugin": "14.0.0-alpha.3",
-    "@angular-eslint/eslint-plugin-template": "14.0.0-alpha.3",
-    "@angular-eslint/schematics": "14.0.0-alpha.3",
-    "@angular-eslint/template-parser": "14.0.0-alpha.3",
+    "@angular-eslint/builder": "14.0.3",
+    "@angular-eslint/eslint-plugin": "14.0.3",
+    "@angular-eslint/eslint-plugin-template": "14.0.3",
+    "@angular-eslint/schematics": "14.0.3",
+    "@angular-eslint/template-parser": "14.0.3",
     "@angular/cli": "~14.1.3",
     "@angular/compiler-cli": "^14.1.3",
     "@compodoc/compodoc": "^1.1.19",
     "@percy/cli": "^1.8.1",
     "@percy/cypress": "^3.1.2",
+    "@schematics/angular": "^14.2.1",
     "@types/jasmine": "~4.0.0",
-    "@typescript-eslint/eslint-plugin": "5.27.1",
-    "@typescript-eslint/parser": "5.27.1",
-    "eslint": "^8.17.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
+    "eslint": "^8.18.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jasmine-core": "~4.1.0",
@@ -80,6 +81,7 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "ng-packagr": "^14.0.0",
     "npm-run-all": "^4.1.5",
+    "prettier": "^2.7.1",
     "typescript": "~4.7.2"
   }
 }
