@@ -1,7 +1,7 @@
-import { defineConfig } from 'cypress'
+import { defineConfig } from 'cypress';
 
 export default defineConfig({
   e2e: {
-    baseUrl: 'http://localhost:4200'
-  }
-})
+    baseUrl: 'http://localhost:4200',
+  },
+});
