@@ -1,11 +1,9 @@
-import { Renderer2 } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ChartComponent } from './chart.component';
 
 describe('ChartComponent', () => {
   let component: ChartComponent;
   let fixture: ComponentFixture<ChartComponent>;
-  let renderer: Renderer2;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -16,43 +14,6 @@ describe('ChartComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(ChartComponent);
     component = fixture.componentInstance;
-    component.unlistenMouseWheel = () => {
-      return;
-    };
-    component.unlistenPointerEnter = () => {
-      return;
-    };
-    component.unlistenPointerLeave = () => {
-      return;
-    };
-    component.unlistenPointerMove = () => {
-      return;
-    };
-    component.unlistenTouchStart = () => {
-      return;
-    };
-    component.dataMarksComponent = {
-      config: { tooltip: { show: false } },
-    } as any;
-  });
-
-  describe('ngOnInit()', () => {
-    beforeEach(() => {
-      spyOn(component as any, 'setAspectRatio');
-      spyOn(component as any, 'createDimensionObservables');
-    });
-
-    it('calls setAspectRatio', () => {
-      component.ngOnInit();
-      expect((component as any).setAspectRatio).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls createDimensionObservables', () => {
-      component.ngOnInit();
-      expect(
-        (component as any).createDimensionObservables
-      ).toHaveBeenCalledTimes(1);
-    });
   });
 
   describe('ngOnChanges()', () => {
@@ -75,10 +36,26 @@ describe('ChartComponent', () => {
     });
   });
 
-  describe('ngAfterContentInit()', () => {
+  describe('ngOnInit()', () => {
     beforeEach(() => {
-      spyOn(component as any, 'setPointerEventListeners');
+      spyOn(component as any, 'setAspectRatio');
+      spyOn(component as any, 'createDimensionObservables');
     });
+
+    it('calls setAspectRatio', () => {
+      component.ngOnInit();
+      expect((component as any).setAspectRatio).toHaveBeenCalledTimes(1);
+    });
+
+    it('calls createDimensionObservables', () => {
+      component.ngOnInit();
+      expect(
+        (component as any).createDimensionObservables
+      ).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('ngAfterContentInit()', () => {
     it('should throw an error if no dataMarks component exists', () => {
       component.dataMarksComponent = undefined;
       expect(() => {
@@ -88,48 +65,20 @@ describe('ChartComponent', () => {
 
     describe('dataMarksComponent is defined', () => {
       it('should not throw an error', () => {
+        component.dataMarksComponent = {} as any;
         expect(() => {
           component.ngAfterContentInit();
         }).not.toThrow();
       });
-
-      it('calls setPointerEventListeners if dataMarks.config.showTooltip is true', () => {
-        component.dataMarksComponent.config = {
-          tooltip: { show: true },
-        } as any;
-        component.ngAfterContentInit();
-        expect(
-          (component as any).setPointerEventListeners
-        ).toHaveBeenCalledTimes(1);
-      });
-
-      it('does not call setPointerEventListeners if dataMarks.config.showTooltip is false', () => {
-        component.dataMarksComponent.config = {
-          tooltip: { show: false },
-        } as any;
-        component.ngAfterContentInit();
-        expect(
-          (component as any).setPointerEventListeners
-        ).not.toHaveBeenCalled();
-      });
     });
   });
 
-  describe('ngAfterViewInit', () => {
-    beforeEach(() => {
-      spyOn(component, 'setTooltipPosition');
-    });
-
-    it('calls setTooltipPosition once if htmlTooltip exists', () => {
-      component.htmlTooltip.exists = true;
-      component.ngAfterViewInit();
-      expect(component.setTooltipPosition).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls setTooltipPosition once if htmlTooltip does not exist', () => {
-      component.htmlTooltip.exists = false;
-      component.ngAfterViewInit();
-      expect(component.setTooltipPosition).toHaveBeenCalledTimes(0);
+  describe('setAspectRatio()', () => {
+    it('sets aspectRatio to the correct value', () => {
+      component.width = 100;
+      component.height = 50;
+      component.setAspectRatio();
+      expect(component.aspectRatio).toBe(2);
     });
   });
 
@@ -153,92 +102,6 @@ describe('ChartComponent', () => {
     it('returns false if chartShouldScale is false', () => {
       component.scaleChartWithContainer = false;
       expect(component.chartShouldScale()).toBe(false);
-    });
-  });
-
-  describe('setAspectRatio()', () => {
-    it('sets aspectRatio to the correct value', () => {
-      component.width = 100;
-      component.height = 50;
-      component.setAspectRatio();
-      expect(component.aspectRatio).toBe(2);
-    });
-  });
-
-  describe('setPointerEventListeners()', () => {
-    let setTouchStartSpy: jasmine.Spy;
-    let setPointerEnterSpy: jasmine.Spy;
-    let setMouseWheelSpy: jasmine.Spy;
-    beforeEach(() => {
-      setTouchStartSpy = spyOn(component as any, 'setTouchStartListener');
-      setPointerEnterSpy = spyOn(component as any, 'setPointerEnterListener');
-      setMouseWheelSpy = spyOn(component as any, 'setMouseWheelListener');
-      component.svgRef = {
-        nativeElement: 'svg',
-      } as any;
-    });
-    it('should call setTouchStartListener once with the correct value', () => {
-      (component as any).setPointerEventListeners();
-      expect(setTouchStartSpy).toHaveBeenCalledOnceWith('svg');
-    });
-    it('should call setPointerEnterListener once with the correct value', () => {
-      (component as any).setPointerEventListeners();
-      expect(setPointerEnterSpy).toHaveBeenCalledOnceWith('svg');
-    });
-    it('should call setMouseWheelListener once with the correct value', () => {
-      (component as any).setPointerEventListeners();
-      expect(setMouseWheelSpy).toHaveBeenCalledOnceWith('svg');
-    });
-  });
-
-  describe('setTooltipPosition()', () => {
-    beforeEach(() => {
-      component.divRef = {
-        nativeElement: {
-          getBoundingClientRect: jasmine
-            .createSpy('getBoundingClientRect')
-            .and.returnValue({ x: 100, y: 200 }) as any,
-        } as any,
-      } as any;
-    });
-
-    describe('if htmlTooltip.exists', () => {
-      beforeEach(() => {
-        component.htmlTooltip.exists = true;
-        component.setTooltipPosition();
-      });
-      it('correctly sets position.top on htmlTooltip', () => {
-        expect(component.htmlTooltip.position.top).toEqual(200);
-      });
-
-      it('correctly sets position.left on htmlTooltip', () => {
-        expect(component.htmlTooltip.position.left).toEqual(100);
-      });
-    });
-
-    describe('if htmlTooltip.exists is falsy', () => {
-      beforeEach(() => {
-        component.htmlTooltip.exists = false;
-        component.setTooltipPosition();
-      });
-      it('does not set position.top on htmlTooltip', () => {
-        expect(component.htmlTooltip.position.top).toEqual(0);
-      });
-      it('does not set position.left on htmlTooltip', () => {
-        expect(component.htmlTooltip.position.left).toEqual(0);
-      });
-    });
-  });
-
-  describe('emitTooltipData()', () => {
-    beforeEach(() => {
-      spyOn(component.tooltipData, 'emit');
-    });
-    it('should call emit on tooltipData once with the correct value', () => {
-      component.emitTooltipData('test' as any);
-      expect(component.tooltipData.emit).toHaveBeenCalledOnceWith(
-        'test' as any
-      );
     });
   });
 });
