@@ -1,16 +1,12 @@
 import {
   AfterContentInit,
-  AfterViewInit,
   ChangeDetectionStrategy,
   Component,
   ContentChild,
   ElementRef,
-  EventEmitter,
   Input,
   OnChanges,
-  OnDestroy,
   OnInit,
-  Output,
   Renderer2,
   SimpleChanges,
   ViewChild,
@@ -29,7 +25,6 @@ import {
 } from 'rxjs';
 import { DataMarks } from '../data-marks/data-marks';
 import { DATA_MARKS } from '../data-marks/data-marks.token';
-import { HtmlTooltipConfig } from '../html-tooltip/html-tooltip.model';
 
 export interface Ranges {
   x: [number, number];
@@ -53,9 +48,7 @@ export interface Dimensions {
   styleUrls: ['./chart.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ChartComponent
-  implements OnInit, OnChanges, AfterViewInit, AfterContentInit, OnDestroy
-{
+export class ChartComponent implements OnInit, OnChanges, AfterContentInit {
   @ContentChild(DATA_MARKS)
   dataMarksComponent: DataMarks;
   @ViewChild('div', { static: true }) divRef: ElementRef<HTMLDivElement>;
@@ -70,14 +63,7 @@ export class ChartComponent
   };
   @Input() scaleChartWithContainer = true;
   @Input() transitionDuration?: number = 250;
-  @Output() tooltipData = new EventEmitter<any>();
-  unlistenPointerEnter: () => void;
-  unlistenPointerMove: () => void;
-  unlistenPointerLeave: () => void;
-  unlistenTouchStart: () => void;
-  unlistenMouseWheel: () => void;
   aspectRatio: number;
-  htmlTooltip: HtmlTooltipConfig = new HtmlTooltipConfig();
   svgDimensions$: Observable<Dimensions>;
   ranges$: Observable<Ranges>;
   heightSubject = new Subject<number>();
@@ -102,22 +88,6 @@ export class ChartComponent
   ngAfterContentInit(): void {
     if (!this.dataMarksComponent) {
       throw new Error('DataMarksComponent not found.');
-    } else if (this.dataMarksComponent.config.tooltip.show) {
-      this.setPointerEventListeners();
-    }
-  }
-
-  ngAfterViewInit(): void {
-    if (this.htmlTooltip.exists) {
-      this.setTooltipPosition();
-    }
-  }
-
-  ngOnDestroy(): void {
-    if (this.dataMarksComponent?.config.tooltip.show) {
-      this.unlistenTouchStart();
-      this.unlistenPointerEnter();
-      this.unlistenMouseWheel();
     }
   }
 
@@ -195,85 +165,5 @@ export class ChartComponent
       this.scaleChartWithContainer &&
       this.divRef.nativeElement.offsetWidth <= this.width
     );
-  }
-
-  private setPointerEventListeners(): void {
-    const el = this.svgRef.nativeElement;
-    this.setTouchStartListener(el);
-    this.setPointerEnterListener(el);
-    this.setMouseWheelListener(el);
-  }
-
-  private setTouchStartListener(el: Element) {
-    this.unlistenTouchStart = this.renderer.listen(
-      el,
-      'touchstart',
-      (event) => {
-        this.onTouchStart(event);
-      }
-    );
-  }
-
-  private onTouchStart(event: TouchEvent): void {
-    event.preventDefault();
-  }
-
-  private setPointerEnterListener(el: Element) {
-    this.unlistenPointerEnter = this.renderer.listen(
-      el,
-      'pointerenter',
-      (event) => {
-        this.onPointerEnter(event, el);
-      }
-    );
-  }
-
-  private onPointerEnter(event: PointerEvent, el: Element): void {
-    this.dataMarksComponent.onPointerEnter(event);
-    this.setPointerMoveListener(el);
-    this.setPointerLeaveListener(el);
-  }
-
-  private setPointerMoveListener(el) {
-    this.unlistenPointerMove = this.renderer.listen(
-      el,
-      'pointermove',
-      (event) => {
-        this.dataMarksComponent.onPointerMove(event);
-      }
-    );
-  }
-
-  private setPointerLeaveListener(el: Element) {
-    this.unlistenPointerLeave = this.renderer.listen(
-      el,
-      'pointerleave',
-      (event) => {
-        this.dataMarksComponent.onPointerLeave(event);
-        this.unlistenPointerMove();
-        this.unlistenPointerLeave();
-      }
-    );
-  }
-
-  private setMouseWheelListener(el: Element) {
-    this.unlistenMouseWheel = this.renderer.listen(el, 'mousewheel', () => {
-      if (this.htmlTooltip.exists) {
-        this.setTooltipPosition();
-      }
-    });
-  }
-
-  setTooltipPosition(): void {
-    if (this.htmlTooltip.exists) {
-      this.htmlTooltip.position.top =
-        this.divRef.nativeElement.getBoundingClientRect().y;
-      this.htmlTooltip.position.left =
-        this.divRef.nativeElement.getBoundingClientRect().x;
-    }
-  }
-
-  emitTooltipData<T>(data: T): void {
-    this.tooltipData.emit(data);
   }
 }
