@@ -20,15 +20,24 @@ export * from './lib/chart/chart.module';
 export * from './lib/data-marks/data-dimension.config';
 export * from './lib/data-marks/data-marks.config';
 export * from './lib/data-marks/data-marks.token';
+export * from './lib/events/chart-svg-event';
+export * from './lib/events/click-event';
+export * from './lib/events/hover-event';
+export * from './lib/events/hover-move-event';
+export * from './lib/events/input-event';
 export * from './lib/geographies/geographies.component';
 export * from './lib/geographies/geographies.config';
 export * from './lib/geographies/geographies.module';
 export * from './lib/grouped-bars/grouped-bars.component';
 export * from './lib/grouped-bars/grouped-bars.config';
 export * from './lib/grouped-bars/grouped-bars.module';
-export * from './lib/html-tooltip/html-tooltip.component';
-export * from './lib/html-tooltip/html-tooltip.model';
+export * from './lib/html-tooltip/html-tooltip.directive';
 export * from './lib/html-tooltip/html-tooltip.module';
+export * from './lib/lines/lines-effect';
+export * from './lib/lines/lines-effects';
+export * from './lib/lines/lines-effects-default-styles.config';
+export * from './lib/lines/lines-hover-move-event.directive';
+export * from './lib/lines/lines-input-event.directive';
 export * from './lib/lines/lines.component';
 export * from './lib/lines/lines.config';
 export * from './lib/lines/lines.module';
